VulnRecon Suite/requirements.txt
python-whois
dnspython
requests
python-nmap
shodan
